+++ bot/api/v1/serializers.py
@@ -18,31 +18,3 @@ class BotOrderSerializer(serializers.ModelSerializer):
         fields = '__all__'
 
 
-class CustomAuthTokenSerializer(serializers.Serializer):
-    email = serializers.CharField(label=_("Email"), write_only=True)
-    password = serializers.CharField(
-        label=_("Password"),
-        style={"input_type": "password"},
-        trim_whitespace=False,
-        write_only=True,
-    )
-    token = serializers.CharField(label=_("Token"), read_only=True)
-
-    def validate(self, attrs):
-        email = attrs.get("email")
-        password = attrs.get("password")
-
-        if email and password:
-            user = authenticate(
-                request=self.context.get("request"),
-                username=email,
-                password=password,
-            )
-
-            if not user or not user.is_verified:
-                raise AuthenticationFailed(
-                    _("Unable to log in. User credentials are invalid or the user is not verified."),
-                    code="authorization")
-
-            attrs["user"] = user
-            return attrs
