+++ bot/api/v1/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 from .views import UserModelViewSet, OrderModelViewSet
-from rest_framework.authtoken.views import obtain_auth_token
+
 """commented for v1 """
 # from .views import UserModelViewSet, EmailModelViewSet, OrderModelViewSet
 
@@ -17,5 +17,4 @@ router.register('BotOrder', OrderModelViewSet, basename='BotOrder')
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('Token/', obtain_auth_token, name='api_token'),
 ]
