+++ bot/api/v1/views.py
@@ -1,4 +1,4 @@
-from .serializers import BotOrderSerializer, BotOrderSerializer, CustomAuthTokenSerializer
+from .serializers import BotOrderSerializer, BotOrderSerializer
 from rest_framework import viewsets
 from ...models import User, Order
 from rest_framework.authtoken.views import ObtainAuthToken
@@ -25,17 +25,5 @@ class OrderModelViewSet(viewsets.ModelViewSet):
     queryset = Order.objects.all()
     serializer_class = BotOrderSerializer
 
-
-class CustomObtainAuthToken(ObtainAuthToken):
-    serializer_class = CustomAuthTokenSerializer
-
-    def post(self, request, *args, **kwargs):
-        serializer = self.serializer_class(
-            data=request.data, context={"request": request}
-        )
-        serializer.is_valid(raise_exception=True)
-        user = serializer.validated_data["user"]
-        token, created = Token.objects.get_or_create(user=user)
-        return Response({"token": token.key, "user_id": user.pk, "email": user.email})
