+++ website/api/v1/views.py
@@ -1,7 +1,8 @@
 from ...models import Configuration, Product, TelegramChannel, Tutorial, ChannelAdmin, User
 from rest_framework.viewsets import ModelViewSet
 from .serializers import ConfigurationSerializer, ProductSerializer, TelegramChannelSerializer, TutorialSerializer, \
-    ChannelAdminSerializer, MessageSerializer, WebsiteUserSerializer
+    ChannelAdminSerializer, MessageSerializer, WebsiteUserSerializer, CustomAuthTokenSerializer
+from rest_framework.authtoken.views import ObtainAuthToken
 
 
 class UserModelViewSet(ModelViewSet):
@@ -37,3 +38,16 @@ class ChannelAdminModelViewSet(ModelViewSet):
 class MessageModelViewSet(ModelViewSet):
     queryset = ChannelAdmin.objects.all()
     serializer_class = MessageSerializer
+
+
+class CustomObtainAuthToken(ObtainAuthToken):
+    serializer_class = CustomAuthTokenSerializer
+
+    def post(self, request, *args, **kwargs):
+        serializer = self.serializer_class(
+            data=request.data, context={"request": request}
+        )
+        serializer.is_valid(raise_exception=True)
+        user = serializer.validated_data["user"]
+        token, created = Token.objects.get_or_create(user=user)
+        return Response({"token": token.key, "user_id": user.pk, "email": user.email})
